feat: fetch default config if no config id is found
