+++ numalogic/udfs/postprocess.py
@@ -11,10 +11,9 @@ from pynumaflow.mapper import Messages, Datum, Message
 
 from numalogic.config import PostprocessFactory, RegistryFactory
 from numalogic.registry import LocalLRUCache
-from numalogic.tools.exceptions import ConfigNotFoundError
 from numalogic.tools.types import redis_client_t, artifact_t
 from numalogic.udfs import NumalogicUDF
-from numalogic.udfs._config import StreamConf, PipelineConf
+from numalogic.udfs._config import PipelineConf
 from numalogic.udfs.entities import StreamPayload, Header, Status, TrainerPayload, OutputPayload
 from numalogic.udfs.tools import _load_artifact
 
@@ -35,13 +34,14 @@ class PostprocessUDF(NumalogicUDF):
         pl_conf: PipelineConf object
     """
 
+    __slots__ = ("registry_conf", "model_registry", "postproc_factory")
+
     def __init__(
         self,
         r_client: redis_client_t,
         pl_conf: Optional[PipelineConf] = None,
     ):
-        super().__init__()
-        self.pl_conf = pl_conf or PipelineConf()
+        super().__init__(pl_conf=pl_conf)
         self.registry_conf = self.pl_conf.registry_conf
         model_registry_cls = RegistryFactory.get_cls(self.registry_conf.name)
         self.model_registry = model_registry_cls(
@@ -54,22 +54,6 @@ class PostprocessUDF(NumalogicUDF):
             ),
         )
         self.postproc_factory = PostprocessFactory()
-
-    def register_conf(self, config_id: str, conf: StreamConf) -> None:
-        """
-        Register config with the UDF.
-
-        Args:
-            config_id: Config ID
-            conf: StreamConf object
-        """
-        self.pl_conf.stream_confs[config_id] = conf
-
-    def get_conf(self, config_id: str) -> StreamConf:
-        try:
-            return self.pl_conf.stream_confs[config_id]
-        except KeyError as err:
-            raise ConfigNotFoundError(f"Config with ID {config_id} not found!") from err
 
     def exec(self, keys: list[str], datum: Datum) -> Messages:
         """
